+++ src/hotspot/share/gc/serial/cardTableRS.hpp
@@ -22,8 +22,8 @@
  *
  */
 
-#ifndef SHARE_GC_SHARED_CARDTABLERS_HPP
-#define SHARE_GC_SHARED_CARDTABLERS_HPP
+#ifndef SHARE_GC_SERIAL_CARDTABLERS_HPP
+#define SHARE_GC_SERIAL_CARDTABLERS_HPP
 
 #include "gc/shared/cardTable.hpp"
 #include "memory/memRegion.hpp"
