+++ src/hotspot/share/gc/serial/defNewGeneration.inline.hpp
@@ -27,7 +27,7 @@
 
 #include "gc/serial/defNewGeneration.hpp"
 
-#include "gc/shared/cardTableRS.hpp"
+#include "gc/serial/cardTableRS.hpp"
 #include "gc/shared/genCollectedHeap.hpp"
 #include "gc/shared/space.inline.hpp"
 #include "oops/access.inline.hpp"
