+++ src/hotspot/share/gc/serial/vmStructs_serial.hpp
@@ -25,6 +25,7 @@
 #ifndef SHARE_GC_SERIAL_VMSTRUCTS_SERIAL_HPP
 #define SHARE_GC_SERIAL_VMSTRUCTS_SERIAL_HPP
 
+#include "gc/serial/cardTableRS.hpp"
 #include "gc/serial/serialHeap.hpp"
 #include "gc/serial/tenuredGeneration.hpp"
 
@@ -69,6 +70,7 @@
   declare_type(TenuredSpace,                 ContiguousSpace)                 \
                                                                               \
   declare_type(DefNewGeneration,             Generation)                      \
+  declare_type(CardTableRS, CardTable)                                        \
                                                                               \
   declare_toplevel_type(TenuredGeneration*)                                   \
   declare_toplevel_type(BlockOffsetSharedArray)                               \
